adding security to API according to role

+++ src/main/java/es/zenith/levelestapi/service/UserService.java
@@ -10,6 +10,11 @@ import es.zenith.levelestapi.mapper.UserMapper;
 import es.zenith.levelestapi.repository.UserRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import jakarta.annotation.PostConstruct;
@@ -41,21 +46,38 @@ public class UserService {
     @Autowired
     private LevelService levelService;
 
+    @Autowired
+    private PasswordEncoder passwordEncoder;
+
 
     @PostConstruct
     public void init() {
         if (userRepository.count()>0) return;
         List<User> users = List.of(
-                new User("Victor","12345","Victor Hugo","OLIVEIRA","victor@gmail.com", List.of("USER", "DEVELOPER") ),
-                new User("Naroa","12345","Naroa","Martín","naroa@gmail.com", List.of("USER", "DEVELOPER") ),
-                new User("David","12345","David","Martín","david@gmail.com", List.of("USER", "DEVELOPER") ),
-                new User("Jaime","12345","Jaime","Ochoa de Alda","jaime@gmail.com", List.of("USER", "DEVELOPER") ),
-                new User("Nerea","12345","Nerea Tindary","Morocho","nerea@gmail.com", List.of("USER", "DEVELOPER") ),
-                new User("Samuel","12345","Samuel","Melían","samuel@gmail.com", List.of("USER", "DEVELOPER") )
+                new User("Victor", passwordEncoder.encode("12345"), "Victor Hugo","OLIVEIRA","victor@gmail.com", List.of("USER", "DEVELOPER") ),
+                new User("Naroa", passwordEncoder.encode("12345"),"Naroa","Martín","naroa@gmail.com", List.of("USER", "DEVELOPER") ),
+                new User("David", passwordEncoder.encode("12345"),"David","Martín","david@gmail.com", List.of("USER", "DEVELOPER") ),
+                new User("Jaime", passwordEncoder.encode("12345"),"Jaime","Ochoa de Alda","jaime@gmail.com", List.of("USER", "DEVELOPER") ),
+                new User("Nerea", passwordEncoder.encode("12345"),"Nerea Tindary","Morocho","nerea@gmail.com", List.of("USER", "DEVELOPER") ),
+                new User("Samuel", passwordEncoder.encode("12345"),"Samuel","Melían","samuel@gmail.com", List.of("USER", "DEVELOPER") )
                 );
         for (User user : users) {
             userRepository.save(user);
         }
+    }
+
+    private User getLoggedUser() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth != null && !(auth instanceof AnonymousAuthenticationToken)) {
+            return userRepository.findByUsername(auth.getName()).get();
+        }
+        return null;
+    }
+
+    private boolean allowedAction(long id){
+        User user = getLoggedUser();
+        if (user == null) return false;
+        return user.getRoles().contains("DEVELOPER") || user.getId().equals(id);
     }
 
     public List<UserSimpleDTO> getUsers() {
@@ -73,15 +95,16 @@ public class UserService {
 
     @Transactional
     public UserDTO saveUser(UserFormDTO userFormDTO) {
-        User user = new  User(userFormDTO.username(), userFormDTO.password(), userFormDTO.name(), userFormDTO.surname(), userFormDTO.email(), List.of("USER"));
+        User user = new  User(userFormDTO.username(), passwordEncoder.encode(userFormDTO.password()), userFormDTO.name(), userFormDTO.surname(), userFormDTO.email(), List.of("USER"));
         return userMapper.toDTO(userRepository.save(user));
     }
 
     @Transactional
     public UserDTO updateUser(long id, UserFormDTO userFormDTO) {
+        if (!allowedAction(id)) throw new AccessDeniedException("Access denied");
         User user = userRepository.findById(id).orElseThrow(() -> new NoSuchElementException("User not found"));
         user.setUsername(userFormDTO.username());
-        user.setPassword(userFormDTO.password());
+        user.setEncodedPassword(userFormDTO.password());
         user.setName(userFormDTO.name());
         user.setSurname(userFormDTO.surname());
         user.setEmail(userFormDTO.email());
@@ -90,6 +113,7 @@ public class UserService {
 
     @Transactional
     public UserDTO deleteUser(long id) {
+        if (!allowedAction(id)) throw new AccessDeniedException("Access denied");
         User user = userRepository.findById(id).orElseThrow(() -> new NoSuchElementException("User not found"));
         userRepository.delete(user);
         return userMapper.toDTO(user);
@@ -106,6 +130,7 @@ public class UserService {
 
     @Transactional
     public LevelDTO addCompletedLevels(long id, long levelId) {
+        if (!allowedAction(id)) throw new AccessDeniedException("Access denied");
         User user = userRepository.findById(id).orElseThrow(() -> new NoSuchElementException("User not found"));
         LevelDTO levelDTO = levelService.getLevel(levelId);
         Level level = levelMapper.toEntity(levelDTO);
@@ -140,7 +165,7 @@ public class UserService {
 
     public UserDTO login(LoginForm loginForm) {
         User user = userRepository.findByEmail(loginForm.getEmail()).orElseThrow(() -> new NoSuchElementException("User not found"));
-        if (user.getPassword().equals(loginForm.getPassword())) {
+        if (user.getEncodedPassword().equals(passwordEncoder.encode(loginForm.getPassword()))) {
             return userMapper.toDTO(user);
         }
         return null;
